+++ index.html
@@ -2,98 +2,30 @@
 <html lang="en">
 
 <head>
-    <meta charset="UTF-8">
-    <title>Tab Menu</title>
-    <style>
-        * {
-            margin: 0;
-            padding: 0;
-            box-sizing: border-box;
-        }
-
-        html {
-            background-color: #0d1117;
-            color: #fff;
-            font-family: sans-serif;
-        }
-
-        p {
-            max-width: 48ch;
-            margin: 1rem;
-        }
-    </style>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Web Components Test</title>
 </head>
 
 <body>
-    <nm-dialog id="non-modal"></nm-dialog>
     <modal-dialog id="modal"></modal-dialog>
-    <button onclick="displayTestModal(false)">Standard Modal</button>
-    <button onclick="displayTestModal(true)">Required Modal</button>
-    <button onclick="displayTestNm()">Non-Modal</button>
 
-    <tab-menu open="Buttons">
-        <div name="Lorem Ipsum">
-            <p>This is some text.</p>
-        </div>
-        <div name="Buttons">
-            <button>Button 1</button><button>Button 2</button><button>Button 3</button>
-        </div>
-        <div name="Picture" disabled>
-            <img src="https://images.pexels.com/photos/2653362/pexels-photo-2653362.jpeg" width="100%"
-                alt="A computer screen displaying CSS code.">
-        </div>
-    </tab-menu>
+    <template id="modal-1">
+        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Architecto dolorum, optio ratione fugit voluptas
+            autem iste iusto corporis totam, facere quae! Tempora veniam laboriosam ab minima voluptates ipsam totam
+            placeat.</p>
+        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Architecto dolorum, optio ratione fugit voluptas
+            autem iste iusto corporis totam, facere quae! Tempora veniam laboriosam ab minima voluptates ipsam totam
+            placeat.</p>
 
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Animi corrupti id necessitatibus. Cupiditate facilis
-        illum necessitatibus possimus, veritatis odit molestias, beatae sapiente a eaque, dolorum architecto tempora aut
-        expedita laudantium.</p>
+        <button class="close">Test</button>
+    </template>
+
+    <template id="modal-2">
+        <h2>This is another dialog.</h2>
+        <button class="close">Test</button>
+    </template>
 
     <script type="module" src="./dialog.js"></script>
-    <script type="module" src="./tabmenu.js"></script>
-    <script>
-        function displayTestModal(required) {
-            let modal = document.getElementById("modal");
-
-            switch (required) {
-                case true:
-                    modal.display(
-                        "Required",
-                        "<p>You can only close this dialog by clicking <button id='close'>here!</button></p>",
-                        true
-                    );
-                    break;
-
-                default:
-                    modal.display(
-                        "Not Required",
-                        "<p>This dialog can be dismissed by clicking anywhere outside it or the 'X' button!</p>",
-                    )
-            }
-        }
-
-        function displayTestNm() {
-            let nm = document.getElementById("non-modal");
-            nm.display(
-                "This dialog can be closed by clicking the 'X' button, but it\
-                will eventually go away on its own.", 5
-            )
-        }
-    </script>
 </body>
 
 </html>
